rename cookieUtils to tokenUtils
